+++ test/OpenTelemetry.Instrumentation.Wcf.Tests/WCF/IServiceContract.cs
@@ -25,9 +25,18 @@ public interface IServiceContract
     [OperationContract]
     Task<ServiceResponse> ExecuteAsync(ServiceRequest request);
 
+    [OperationContract]
+    ServiceResponse ExecuteSynchronous(ServiceRequest request);
+
     [OperationContract(Action = "")]
     Task<ServiceResponse> ExecuteWithEmptyActionNameAsync(ServiceRequest request);
 
     [OperationContract(IsOneWay = true)]
     void ExecuteWithOneWay(ServiceRequest request);
+
+    [OperationContract]
+    void ErrorSynchronous();
+
+    [OperationContract]
+    Task ErrorAsync();
 }