+++ test/OpenTelemetry.Instrumentation.Wcf.Tests/WCF/ServiceClient.cs
@@ -30,9 +30,18 @@ public class ServiceClient : ClientBase<IServiceContract>, IServiceContract
     public Task<ServiceResponse> ExecuteAsync(ServiceRequest request)
         => this.Channel.ExecuteAsync(request);
 
+    public ServiceResponse ExecuteSynchronous(ServiceRequest request)
+        => this.Channel.ExecuteSynchronous(request);
+
     public Task<ServiceResponse> ExecuteWithEmptyActionNameAsync(ServiceRequest request)
         => this.Channel.ExecuteWithEmptyActionNameAsync(request);
 
     public void ExecuteWithOneWay(ServiceRequest request)
         => this.Channel.ExecuteWithOneWay(request);
+
+    public Task ErrorAsync()
+        => this.Channel.ErrorAsync();
+
+    public void ErrorSynchronous()
+        => this.Channel.ErrorSynchronous();
 }