+++ src/OpenTelemetry.Instrumentation.Wcf/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Update OpenTelemetry SDK version to `1.5.1`.
   ([#1255](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1255))
+* Client instrumentation implementation moved to lower-level `BindingElement`.
+  ([#1247](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1247))
 
 ## 1.0.0-rc.10